Factor fft sampling code out into new file
